Change search movie route validation, sanitization and results, and update tests
Now:

The query can be empty. In this case, there is a default response without make a request to tmdb api.

This route receive enconded query and decoded it.

The query is returned decoded in the new field query.

The prev_page and next_page fields have the query enconded when these are not empty.

Update jest tests.